requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
cryptography==42.0.2
pytest==7.4.3
httpx==0.26.0
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0 